Add ability to use a custom install script
A custom install script can be used instead of installing Docker. If the
driver used to create the machine has userdata, then the userdata must
be in #! or #cloud-config format and the install script contents will be
added to the userdata. If the driver does not have userdata, then the
install script will be run on the machine via ssh.

Also, since Docker is not being installed, then the certificate
information is not created for the machine.