Set proper exit code on pod exit
If the rancher-machine image is run in a pod, then the exit code of the
pod was not being properly set if the rancher-machine call failed (set
to 0 even though rancher-machine exited with code 1). The changes to the
entrypoint script properly sets the exit code if the rancher-machine
command fails while still passing the ouput of the command to tee.