Allow machine to trust Rancher certs
The machine pod was not able to trust the Rancher certificates passed to the job
for downloading the drivers. This change will run c_rehash when the machine pod
starts up to allow the curl commands to work.

In addition to using the c_rehash command, the base image is changed to SLE BCI to
be consistent with Rancher.