Separate path and command
With the introduction of Windows support, the path to and the command to run the
install script is different in Linux and in Windows. An effort was made to do
this, but he command and path were combined. This caused a problem because the
file was put in a very strange place and the file could not be found when the
call to run the command was made.

This change separates the command from the path to avoid this issue.