Link valid certs for cert rehasing
A previous change called c_rehash before downloading a driver in cae the driver
was downloaded from Rancher and Rancher was using a self-signed certificate.
This caused a problem when the driver was download from a source outside of
Rancher because the valid certificates were no longer trusted.

This change will link the trusted certificates into `SSL_CERT_DIR`, if it is set
so they can be included.